Add the "Xubuntu Fifteen" theme to the repository.
This theme is the new Xubuntu website default theme. It is based on "Xubuntu Fourteen", but since the visual changes are so dramatic, it should be a separate theme. Amongst other things, the new theme introduces responsive design for small-resolution devices.
The backend side of the two themes isn't that different; that said, a code review should be relatively simple.

This commit reverts the "Xubuntu Fourteen" theme to the state it was in revision 50; the last version that was pushed to production.